+++ src/register.tsx
@@ -7,7 +7,7 @@ import { AddonPanel } from '@storybook/components';
 import Panel from './components/Panel/Panel';
 import SettingsProvider from './contexts/SettingsProvider';
 
-addons.register('storybook-addon-theme-playground', api => {
+addons.register('storybook-addon-theme-playground', (api) => {
   addons.addPanel('storybook-addon-theme-playground/panel', {
     title: 'Theme Playground',
     render: ({ active, key }) => (
@@ -16,6 +16,6 @@ addons.register('storybook-addon-theme-playground', api => {
           <Panel />
         </SettingsProvider>
       </AddonPanel>
-    )
+    ),
   });
 });