+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.1]- 2020-06-27
+
+### Added
+
+- `lodash.startcase` dependency
+
+### Fixed
+
+- [PR](https://github.com/jeslage/storybook-addon-theme-playground/pull/5): Bump websocket-extensions from 0.1.3 to 0.1.4
+
 ## [1.3.0]- 2020-03-24
 
 ### Added
@@ -78,7 +88,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.3.0...develop
+[unreleased]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.3.1...develop
+[1.3.1]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.3.0...1.3.1
 [1.3.0]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.2.2...1.3.0
 [1.2.2]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.2.0...v1.2.1